Unit tests have been rewritten and added README.md has been changed to be more precise.

+++ ds-image-analysis-ws/src/main/java/dk/kb/api/webservice/DHash.java
@@ -45,36 +45,47 @@ public class DHash extends ImageHash {
     protected List<String> generateJSON() {
         List<String> reply = new LinkedList<>();
         try {
-            FileUtils.copyURLToFile(getImgURL(), getHashPath().toFile());
-            File jsonFile = Files.createTempFile("DHash", ".JSON").toFile();
-            jsonFile.deleteOnExit();
-            JsonGenerator jsonGenerator = new JsonFactory()
-                    .createGenerator(new FileOutputStream(jsonFile));
-            //for pretty printing
-            jsonGenerator.setPrettyPrinter(new DefaultPrettyPrinter());
-            jsonGenerator.writeStartObject(); // start root object
+            if ((getImgURL() != null) && (!getImgURL().toString().isEmpty())) {
+                FileUtils.copyURLToFile(getImgURL(), getHashPath().toFile());
+                File jsonFile = Files.createTempFile("DHash", ".JSON").toFile();
+                jsonFile.deleteOnExit();
+                JsonGenerator jsonGenerator = new JsonFactory()
+                        .createGenerator(new FileOutputStream(jsonFile));
+                //for pretty printing
+                jsonGenerator.setPrettyPrinter(new DefaultPrettyPrinter());
+                jsonGenerator.writeStartObject(); // start root object
 
-            jsonGenerator.writeStringField("Algorithm", "Difference Hash");
-            jsonGenerator.writeStringField("Image URL", getImgURL().toString());
+                jsonGenerator.writeStringField("Algorithm", "Difference Hash");
+                jsonGenerator.writeStringField("Image URL", getImgURL().toString());
 
-            setPrecision(DifferenceHash.Precision.Simple);
-            hashValue(jsonGenerator);
+                if ((getStart() < 1) || (getEnd() < 1) || (getStart() > getEnd())) {
+                    reply.add("Both start and end have to be positive and start value has to be less or equal to end value");
+                    return reply;
+                }
+                setPrecision(DifferenceHash.Precision.Simple);
+                hashValue(jsonGenerator);
 
-            setPrecision(DifferenceHash.Precision.Double);
-            hashValue(jsonGenerator);
+                setPrecision(DifferenceHash.Precision.Double);
+                hashValue(jsonGenerator);
 
-            setPrecision(DifferenceHash.Precision.Triple);
-            hashValue(jsonGenerator);
-            jsonGenerator.writeEndObject();
+                setPrecision(DifferenceHash.Precision.Triple);
+                hashValue(jsonGenerator);
+                jsonGenerator.writeEndObject();
 
-            jsonGenerator.flush();
-            jsonGenerator.close();
+                jsonGenerator.flush();
+                jsonGenerator.close();
 
-            reply = Files.readAllLines(jsonFile.toPath());
-
+                reply = Files.readAllLines(jsonFile.toPath());
+            }
+            else {
+                reply.add("The URL has to defined");
+            }
         } catch (IOException e) {
             reply.add("A file error appeared.");
             log.error("A file error appeared", e);
+        } catch (Exception ex) {
+            reply.add("An unhandled exception happened");
+            log.error("An unhandled exception happened", ex);
         }
         return reply;
     }
